New package to api and db tests
