+++ niffler-e-2-e-tests-hw/src/test/java/guru/qa/niffler/test/fake/api/isolation/IsolatedThirdTests.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.test.api.isolation;
+package guru.qa.niffler.test.fake.api.isolation;
 
 import guru.qa.niffler.service.db.impl.jdbc.UserdataDbClientJdbc;
 import org.junit.jupiter.api.Assertions;
