+++ niffler-e-2-e-tests-hw/src/test/java/guru/qa/niffler/test/fake/db/jdbc/UserdataJdbcTest.java
@@ -1,7 +1,7 @@
-package guru.qa.niffler.test.db.hibernate;
+package guru.qa.niffler.test.fake.db.jdbc;
 
 import guru.qa.niffler.service.UserdataClient;
-import guru.qa.niffler.service.db.impl.hibernate.UserdataDbClientHibernate;
+import guru.qa.niffler.service.db.impl.jdbc.UserdataDbClientJdbc;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -11,9 +11,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Slf4j
-class UserdataHibernateTest {
+class UserdataJdbcTest {
 
-    UserdataClient userdataClient = new UserdataDbClientHibernate();
+    UserdataClient userdataClient = new UserdataDbClientJdbc();
 
     @Test
     void shouldCreateNewUserTest() {
