+++ niffler-e-2-e-tests-hw/src/test/java/guru/qa/niffler/test/fake/db/springJdbc/UserdataSpringJdbcTest.java
@@ -1,7 +1,7 @@
-package guru.qa.niffler.test.db.jdbc;
+package guru.qa.niffler.test.fake.db.springJdbc;
 
 import guru.qa.niffler.service.UserdataClient;
-import guru.qa.niffler.service.db.impl.jdbc.UserdataDbClientJdbc;
+import guru.qa.niffler.service.db.impl.springJdbc.UserdataDbClientSpringJdbc;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -11,9 +11,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Slf4j
-class UserdataJdbcTest {
+class UserdataSpringJdbcTest {
 
-    UserdataClient userdataClient = new UserdataDbClientJdbc();
+    UserdataClient userdataClient = new UserdataDbClientSpringJdbc();
 
     @Test
     void shouldCreateNewUserTest() {
